+++ rapidhash/src/fast.rs
@@ -2,12 +2,12 @@
 //!
 //! This is a specific instantiation of the [rapidhash::inner] module with the following settings:
 //! - `AVALANCHE` is disabled.
-//! - `FNV` is enabled.
+//! - `SPONGE` is enabled.
 //! - `COMPACT` is disabled.
 //! - `PROTECTED` is disabled.
 
 const AVALANCHE: bool = false;
-const FNV: bool = true;
+const SPONGE: bool = true;
 const COMPACT: bool = false;
 const PROTECTED: bool = false;
 
@@ -19,50 +19,50 @@ use crate::inner;
 ///
 /// This is an alias for [inner::RapidHasher] with the following settings:
 /// - `AVALANCHE` is disabled.
-/// - `FNV` is enabled.
+/// - `SPONGE` is enabled.
 /// - `COMPACT` is disabled.
 /// - `PROTECTED` is disabled.
 ///
 /// Use [crate::quality::RapidHasher] for a higher quality hash output where necessary.
-pub type RapidHasher = inner::RapidHasher<AVALANCHE, FNV, COMPACT, PROTECTED>;
+pub type RapidHasher = inner::RapidHasher<AVALANCHE, SPONGE, COMPACT, PROTECTED>;
 
 /// A [std::hash::BuildHasher] trait compatible hasher that uses the [RapidHasher] algorithm.
 ///
 /// This is an alias for [inner::RapidBuildHasher] with the following settings:
 /// - `AVALANCHE` is disabled.
-/// - `FNV` is enabled.
+/// - `SPONGE` is enabled.
 /// - `COMPACT` is disabled.
 /// - `PROTECTED` is disabled.
 ///
 /// Note there that [crate::RapidRandomState] should be used
-pub type RapidBuildHasher = inner::RapidBuildHasher<AVALANCHE, FNV, COMPACT, PROTECTED>;
+pub type RapidBuildHasher = inner::RapidBuildHasher<AVALANCHE, SPONGE, COMPACT, PROTECTED>;
 
 /// A rapidhash equivalent to [std::hash::RandomState] that uses a random seed and secrets.
-pub type RandomState = inner::RandomState<AVALANCHE, FNV, COMPACT, PROTECTED>;
+pub type RandomState = inner::RandomState<AVALANCHE, SPONGE, COMPACT, PROTECTED>;
 
 /// A [std::collections::HashMap] that uses the [RapidHasher] hash.
 ///
 /// This is an alias for [inner::RapidHashMap] with the following settings:
 /// - `AVALANCHE` is disabled.
-/// - `FNV` is enabled.
+/// - `SPONGE` is enabled.
 /// - `COMPACT` is disabled.
 /// - `PROTECTED` is disabled.
 ///
 /// Use [crate::quality::RapidHashMap] where higher hash collision resistance is required.
 #[cfg(any(feature = "std", docsrs))]
-pub type RapidHashMap<K, V> = inner::RapidHashMap<K, V, AVALANCHE, FNV, COMPACT, PROTECTED>;
+pub type RapidHashMap<K, V> = inner::RapidHashMap<K, V, AVALANCHE, SPONGE, COMPACT, PROTECTED>;
 
 /// A [std::collections::HashSet] that uses the [RapidHasher] hash.
 ///
 /// This is an alias for [inner::RapidHashSet] with the following settings:
 /// - `AVALANCHE` is disabled.
-/// - `FNV` is enabled.
+/// - `SPONGE` is enabled.
 /// - `COMPACT` is disabled.
 /// - `PROTECTED` is disabled.
 ///
 /// Use [crate::quality::RapidHashSet] where higher hash collision resistance is required.
 #[cfg(any(feature = "std", docsrs))]
-pub type RapidHashSet<K> = inner::RapidHashSet<K, AVALANCHE, FNV, COMPACT, PROTECTED>;
+pub type RapidHashSet<K> = inner::RapidHashSet<K, AVALANCHE, SPONGE, COMPACT, PROTECTED>;
 
 #[cfg(any(feature = "std", docsrs))]
 pub use inner::HashMapExt;