+++ rapidhash/src/inner/collections.rs
@@ -13,7 +13,7 @@ use crate::inner::RandomState;
 /// let mut map = RapidHashMap::with_capacity_and_hasher(10, Default::default());
 /// map.insert(42, "the answer");
 /// ```
-pub type RapidHashMap<K, V, const AVALANCHE: bool, const FNV: bool, const COMPACT: bool = false, const PROTECTED: bool = false> = std::collections::HashMap<K, V, RandomState<AVALANCHE, FNV, COMPACT, PROTECTED>>;
+pub type RapidHashMap<K, V, const AVALANCHE: bool, const SPONGE: bool, const COMPACT: bool = false, const PROTECTED: bool = false> = std::collections::HashMap<K, V, RandomState<AVALANCHE, SPONGE, COMPACT, PROTECTED>>;
 
 /// A [std::collections::HashSet] type that uses the [RapidBuildHasher] hasher.
 ///
@@ -28,10 +28,10 @@ pub type RapidHashMap<K, V, const AVALANCHE: bool, const FNV: bool, const COMPAC
 /// let mut set = RapidHashSet::with_capacity_and_hasher(10, Default::default());
 /// set.insert("the answer");
 /// ```
-pub type RapidHashSet<K, const AVALANCHE: bool, const FNV: bool, const COMPACT: bool = false, const PROTECTED: bool = false> = std::collections::HashSet<K, RandomState<AVALANCHE, FNV, COMPACT, PROTECTED>>;
+pub type RapidHashSet<K, const AVALANCHE: bool, const SPONGE: bool, const COMPACT: bool = false, const PROTECTED: bool = false> = std::collections::HashSet<K, RandomState<AVALANCHE, SPONGE, COMPACT, PROTECTED>>;
 
 /// A trait for creating a `RapidHashMap` with a specified capacity and hasher.
-pub trait HashMapExt<const AVALANCHE: bool, const FNV: bool, const COMPACT: bool = false, const PROTECTED: bool = false> {
+pub trait HashMapExt<const AVALANCHE: bool, const SPONGE: bool, const COMPACT: bool = false, const PROTECTED: bool = false> {
     /// Create a new `RapidHashMap` with the default capacity and hasher.
     fn new() -> Self;
 
@@ -39,18 +39,18 @@ pub trait HashMapExt<const AVALANCHE: bool, const FNV: bool, const COMPACT: bool
     fn with_capacity(capacity: usize) -> Self;
 }
 
-impl<K, V, const AVALANCHE: bool, const FNV: bool, const COMPACT: bool, const PROTECTED: bool> HashMapExt<AVALANCHE, FNV, COMPACT, PROTECTED> for RapidHashMap<K, V, AVALANCHE, FNV, COMPACT, PROTECTED> {
+impl<K, V, const AVALANCHE: bool, const SPONGE: bool, const COMPACT: bool, const PROTECTED: bool> HashMapExt<AVALANCHE, SPONGE, COMPACT, PROTECTED> for RapidHashMap<K, V, AVALANCHE, SPONGE, COMPACT, PROTECTED> {
     fn new() -> Self {
         RapidHashMap::default()
     }
 
     fn with_capacity(capacity: usize) -> Self {
-        RapidHashMap::with_capacity_and_hasher(capacity, RandomState::<AVALANCHE, FNV, COMPACT, PROTECTED>::default())
+        RapidHashMap::with_capacity_and_hasher(capacity, RandomState::<AVALANCHE, SPONGE, COMPACT, PROTECTED>::default())
     }
 }
 
 /// A trait for creating a `RapidHashSet` with a specified capacity and hasher.
-pub trait HashSetExt<const AVALANCHE: bool, const FNV: bool, const COMPACT: bool = false, const PROTECTED: bool = false> {
+pub trait HashSetExt<const AVALANCHE: bool, const SPONGE: bool, const COMPACT: bool = false, const PROTECTED: bool = false> {
     /// Create a new `RapidHashSet` with the default capacity and hasher.
     fn new() -> Self;
 
@@ -58,12 +58,12 @@ pub trait HashSetExt<const AVALANCHE: bool, const FNV: bool, const COMPACT: bool
     fn with_capacity(capacity: usize) -> Self;
 }
 
-impl<K, const AVALANCHE: bool, const FNV: bool, const COMPACT: bool, const PROTECTED: bool> HashSetExt<AVALANCHE, FNV, COMPACT, PROTECTED> for RapidHashSet<K, AVALANCHE, FNV, COMPACT, PROTECTED> {
+impl<K, const AVALANCHE: bool, const SPONGE: bool, const COMPACT: bool, const PROTECTED: bool> HashSetExt<AVALANCHE, SPONGE, COMPACT, PROTECTED> for RapidHashSet<K, AVALANCHE, SPONGE, COMPACT, PROTECTED> {
     fn new() -> Self {
         RapidHashSet::default()
     }
 
     fn with_capacity(capacity: usize) -> Self {
-        RapidHashSet::with_capacity_and_hasher(capacity, RandomState::<AVALANCHE, FNV, COMPACT, PROTECTED>::default())
+        RapidHashSet::with_capacity_and_hasher(capacity, RandomState::<AVALANCHE, SPONGE, COMPACT, PROTECTED>::default())
     }
 }