+++ rapidhash/src/inner/mod.rs
@@ -8,9 +8,10 @@
 //! Each structure may have the compile time const generics:
 //! - `AVALANCHE`: Whether to use a final avalanche mix step, required to pass SMHasher3. This
 //!   option changes the hash output. Enabled on [rapidhash::quality], disabled on [rapidhash::fast].
-//! - `FNV`: Allow RapidHasher to use FNV when hashing integer types: hashing ints will be twice as
-//!   fast but the hash quality will be lower. This changes the hash output when hashing integers.
-//!   Disabled on [rapidhash::quality], enabled on [rapidhash::fast].
+//! - `SPONGE`: Allow RapidHasher to cache integers into a 128-bit buffer to perform a single
+//!   folded multiply step on the entire buffer. If disabled, a mix step is performed on each
+//!   individual integer. This changes the hash output when hashing integers. Enabled on both
+//!   [rapidhash::quality] and [rapidhash::fast].
 //! - `COMPACT`: Reduce the code size of the hasher by preventing manually unrolled loops. This does
 //!   _not_ affect the hash output. Disabled on both [rapidhash::quality] and [rapidhash::fast].
 //! - `PROTECTED`: When performing the folded multiply mix step, XOR the a and b back into their
@@ -52,8 +53,8 @@ mod tests {
     use rand::Rng;
     use super::{rapidhash_rs, rapidhash_rs_seeded, RAPID_SEED};
 
-    type RapidHasher = super::RapidHasher<false, false>;
-    type RapidBuildHasher = super::RapidBuildHasher<false, false>;
+    type RapidHasher = super::RapidHasher<true, true, true>;
+    type RapidBuildHasher = super::RapidBuildHasher<true, true, true>;
 
     #[derive(Hash)]
     struct Object {
@@ -66,12 +67,12 @@ mod tests {
         let object = Object { bytes: b"hello world".to_vec() };
         let mut hasher = RapidHasher::default();
         object.hash(&mut hasher);
-        assert_eq!(hasher.finish(), 16959177219018390528);
+        assert_eq!(hasher.finish(), 1790036888308448300);
 
         let mut hasher = RapidHasher::default();
         hasher.write_usize(b"hello world".len());
         hasher.write(b"hello world");
-        assert_eq!(hasher.finish(), 16959177219018390528);
+        assert_eq!(hasher.finish(), 1790036888308448300);
     }
 
     /// Check RapidHasher is equivalent to the raw rapidhash for a single byte stream.