+++ rapidhash/src/inner/state.rs
@@ -19,12 +19,12 @@ use crate::inner::RapidHasher;
 /// map.insert(42, "the answer");
 /// ```
 #[derive(Copy, Clone, Eq, PartialEq)]
-pub struct RandomState<const AVALANCHE: bool, const FNV: bool, const COMPACT: bool, const PROTECTED: bool> {
+pub struct RandomState<const AVALANCHE: bool, const SPONGE: bool, const COMPACT: bool, const PROTECTED: bool> {
     seed: u64,
     secrets: &'static [u64; 7],
 }
 
-impl<const AVALANCHE: bool, const FNV: bool, const COMPACT: bool, const PROTECTED: bool> RandomState<AVALANCHE, FNV, COMPACT, PROTECTED> {
+impl<const AVALANCHE: bool, const SPONGE: bool, const COMPACT: bool, const PROTECTED: bool> RandomState<AVALANCHE, SPONGE, COMPACT, PROTECTED> {
     /// Create a new random state with a random seed.
     ///
     /// With the `rand` feature enabled, this will use [rand::random] to initialise the seed.
@@ -61,15 +61,15 @@ impl<const AVALANCHE: bool, const FNV: bool, const COMPACT: bool, const PROTECTE
 }
 
 #[cfg(target_has_atomic = "ptr")]
-impl<const AVALANCHE: bool, const FNV: bool, const COMPACT: bool, const PROTECTED: bool> Default for RandomState<AVALANCHE, FNV, COMPACT, PROTECTED> {
+impl<const AVALANCHE: bool, const SPONGE: bool, const COMPACT: bool, const PROTECTED: bool> Default for RandomState<AVALANCHE, SPONGE, COMPACT, PROTECTED> {
     #[inline]
     fn default() -> Self {
         Self::new()
     }
 }
 
-impl<const AVALANCHE: bool, const FNV: bool, const COMPACT: bool, const PROTECTED: bool>  BuildHasher for RandomState<AVALANCHE, FNV, COMPACT, PROTECTED> {
-    type Hasher = RapidHasher<AVALANCHE, FNV, COMPACT, PROTECTED>;
+impl<const AVALANCHE: bool, const SPONGE: bool, const COMPACT: bool, const PROTECTED: bool>  BuildHasher for RandomState<AVALANCHE, SPONGE, COMPACT, PROTECTED> {
+    type Hasher = RapidHasher<AVALANCHE, SPONGE, COMPACT, PROTECTED>;
 
     #[inline(always)]
     fn build_hasher(&self) -> Self::Hasher {