+++ rapidhash/src/quality.rs
@@ -2,13 +2,13 @@
 //!
 //! This is a specific instantiation of the [rapidhash::inner] module with the following settings:
 //! - `AVALANCHE` is enabled.
-//! - `FNV` is disabled.
+//! - `SPONGE` is enabled.
 //! - `COMPACT` is disabled.
 //! - `PROTECTED` is disabled.
 
 const AVALANCHE: bool = true;
-const FNV: bool = true;
-const COMPACT: bool = true;
+const SPONGE: bool = true;
+const COMPACT: bool = false;
 const PROTECTED: bool = false;
 
 use crate::inner;
@@ -17,51 +17,51 @@ use crate::inner;
 ///
 /// This is an alias for [inner::RapidHasher] with the following settings:
 /// - `AVALANCHE` is enabled.
-/// - `FNV` is disabled.
+/// - `SPONGE` is enabled.
 /// - `COMPACT` is disabled.
 /// - `PROTECTED` is disabled.
 ///
 /// Use [crate::fast::RapidHasher] for a lower quality but faster hash output where desirable.
-pub type RapidHasher = inner::RapidHasher<AVALANCHE, FNV, COMPACT, PROTECTED>;
+pub type RapidHasher = inner::RapidHasher<AVALANCHE, SPONGE, COMPACT, PROTECTED>;
 
 /// A [std::hash::BuildHasher] inspired by [rapidhash::v3::rapidhash_v3] with a focus on hash quality.
 ///
 /// This is an alias for [inner::RapidBuildHasher] with the following settings:
 /// - `AVALANCHE` is enabled.
-/// - `FNV` is disabled.
+/// - `SPONGE` is enabled.
 /// - `COMPACT` is disabled.
 /// - `PROTECTED` is disabled.
 ///
 /// Use [crate::fast::RapidHasher] for a lower quality but faster hash output where desirable.
-pub type RapidBuildHasher = inner::RapidBuildHasher<AVALANCHE, FNV, COMPACT, PROTECTED>;
+pub type RapidBuildHasher = inner::RapidBuildHasher<AVALANCHE, SPONGE, COMPACT, PROTECTED>;
 
 /// A rapidhash equivalent to [std::hash::RandomState] that uses a random seed and secrets for
 /// minimal DoS resistance.
-pub type RandomState = inner::RandomState<AVALANCHE, FNV, COMPACT, PROTECTED>;
+pub type RandomState = inner::RandomState<AVALANCHE, SPONGE, COMPACT, PROTECTED>;
 
 /// A [std::collections::HashMap] that uses the [RapidHasher] hash.
 ///
 /// This is an alias for [inner::RapidHashMap] with the following settings:
 /// - `AVALANCHE` is enabled.
-/// - `FNV` is disabled.
+/// - `SPONGE` is enabled.
 /// - `COMPACT` is disabled.
 /// - `PROTECTED` is disabled.
 ///
 /// Use [crate::fast::RapidHashMap] where lower hash collision resistance is acceptable.
 #[cfg(any(feature = "std", docsrs))]
-pub type RapidHashMap<K, V> = inner::RapidHashMap<K, V, AVALANCHE, FNV, COMPACT, PROTECTED>;
+pub type RapidHashMap<K, V> = inner::RapidHashMap<K, V, AVALANCHE, SPONGE, COMPACT, PROTECTED>;
 
 /// A [std::collections::HashSet] that uses the [RapidHasher] hash.
 ///
 /// This is an alias for [inner::RapidHashSet] with the following settings:
 /// - `AVALANCHE` is enabled.
-/// - `FNV` is disabled.
+/// - `SPONGE` is enabled.
 /// - `COMPACT` is disabled.
 /// - `PROTECTED` is disabled.
 ///
 /// Use [crate::fast::RapidHashSet] where lower hash collision resistance is acceptable.
 #[cfg(any(feature = "std", docsrs))]
-pub type RapidHashSet<K> = inner::RapidHashSet<K, AVALANCHE, FNV, COMPACT, PROTECTED>;
+pub type RapidHashSet<K> = inner::RapidHashSet<K, AVALANCHE, SPONGE, COMPACT, PROTECTED>;
 
 #[cfg(any(feature = "std", docsrs))]
 pub use inner::HashMapExt;